move timedelta test_astype test
